backend/requirements.txt
fastapi
uvicorn
anthropic
pymupdf
pillow
numpy
httpx
python-dotenv
python-multipart
scikit-learn